Working --dst-direct
This makes it possible to make backup and deliver it to remote site, without
writing temporary files on localhost. Thanks to this machine that makes the
backup uses less I/O, and the backup is faster because now we do "create
tarball" and "deliver tarball to remote" at the same time.

This functionality requires some heavy wizardry with shell, so we need that
you have bash available for it.

In process make all local destinations deliver in parallel.