archive-queue check for omnipitr-monitor
This check lets you see if archiving is lagging. Regardless of source of
the problem if the archive-queue check will return value larger than
1 - it means that archiving lags. Usually returned value should be 0.
Sometimes, if you happen to run omnipitr-monitor *while* archiving is in
process - it might return 1. Values larger than that mean that some
segment(s) are waiting in queue.