Fix bug #23
https://github.com/omniti-labs/omnipitr/issues/23

The problem was that when running exec() on a binary that doesn't exist
(bad --rsync-path), exec actually *returns* - which it doesn't when it
succeeds.

Upon returning, it then exits normally, and on exit - destructors are
called, which remove temp files - including the ones with status.

Change detects unsuccessful exec()'s, logs message to stderr (well, to
file), and exits in such a way that destructors are not called.