First set of changes for v 2.0.0
2.0 will work with PostgreSQL 10 and newer, but will *not* work with older
PostgreSQLs.

This is because of change of location of xlog/wal files - previously it
was in pg_xlog/ directory, now it's in pg_wal.

Since the change will not get rolled back, and there is not much
development in OmniPITR, I figured it's simpler to simply provide new
version of OmniPITR that will (should) work with new directory layout of
PostgreSQL 10+, instead of trying to make single OmniPITR version that
works with any Pg version there is.